requirements.txt
scipy==1.5.2
accelerate==0.20.0
easydict==1.11
lpips==0.1.4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
opencv-python==4.8.1.78
numpy==1.26.0
loguru==0.7.2
imageio==2.35.1
pillow==10.4.0